Extracted RealtimeLocker from QSynth for reuse elsewhere
